refactor: added the size block to the base singly linked list struct
This makes the CountNodes function more efficient, as it does not need to traverse the entire list. The appropriate functions have been adapted to support this new style.

+++ include/singly_linked_list.h
@@ -14,7 +14,7 @@ extern "C" {
 #endif
 
 typedef struct SLLNode {
-  int size;             // The total of nodes in the list.
+  size_t size;          // The total of nodes in the list. Autoincrement/decrement
   struct SLLNode *next; // Next element in the list.
 } SLLNode;
